% Generated by roxygen2 (4.0.0): do not edit by hand
\name{GENwincox}
\alias{GENwincox}
\title{Fit coxph genetic analysis on windows.}
\usage{
GENwincox(outcome, covariates, strata, kind, ifile, ofile, pd = "", wd = "")
}
\arguments{
\item{outcome}{A Dataframe/matrix. the first column is the time, and the second column is the status.}

\item{covariates}{vector/Dataframe corresponding to confounding factors. Default
(no covariates) is NULL}

\item{strata}{vector corresponding to stratification variable.Default is no strata.}

\item{ifile}{String corresponding to the name of the plink bfile (no extension)}

\item{ofile}{String corresponding to the name of the resulting files}

\item{pd}{directory of plink. Use only if plink is not in your path}

\item{wd}{directory that all files are located in. Is by default set to the current R working directory}
}
\description{
This function uses R so probably not a good idea to use it on a GWAS. It calls coxph directly from C
so it will be an order of magnitute faster than calling coxph from R.
}

