% Generated by roxygen2 (4.0.0): do not edit by hand
\name{initiate_QC}
\alias{initiate_QC}
\title{Initialize GWAS Quality Control}
\usage{
initiate_QC(ifile, pd = "")
}
\arguments{
\item{pd}{directory where plink is. Do not need to specify if plink is in your path
or working directory}
}
\description{
Looks in current directory for dirty.bed dirty.fam dirty.bin.
Reads these files to get initial data. Saves new files as
GENmatic.bed/fam/bin so that we do not modify origional data
Sets up 2 secret global variables for the rest of the QC process.
This function takes no paramaters
}
\keyword{GWAS}

