% Generated by roxygen2 (4.0.0): do not edit by hand
\name{make_covar_pheno}
\alias{make_covar_pheno}
\title{Make covariate file and phenotype file}
\usage{
make_covar_pheno(type, id, outcome, covar = NULL, strata = NULL,
  cname = "covar", pname = "pheno", wd = "", call = F)
}
\arguments{
\item{type}{What kind of model you want to make the files for. Currently can use coxph, linear, logistic}

\item{id}{datframe corresponding to the 2 column id used in the .fam file}

\item{outcomedataframe}{of covariates wanted in the model. Do not include intercept term. Make sure all factors are actually factors}

\item{strata}{vector of stratification status. Only applicable for coxph}

\item{cname}{name of outputed covar file. Default is covar}

\item{pname}{name of outputed phenotype file. Default is pheno}

\item{wd}{name of directory (including trailing slash) where the outputed files will go. By default files go to the current working directory}
}
\description{
Make covariate file and phenotype file. These files can be made seperate if necessary, but will always be made via GENfit call
}

