% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plinkfit}
\alias{plinkfit}
\title{Use plink to fit models}
\usage{
plinkfit(ifile, ofile = NULL, type, covar, pd, wd)
}
\arguments{
\item{ifile}{String corresponding to the name of bed/bim/fam plink files.}

\item{ofile}{String corresponding to the name of the resulting files. Default is ifile}

\item{type}{type of analysis (coxph, linear, logistic)}

\item{pd}{Directory that plink is located in. Only specify if it is not in your path}

\item{wd}{Directory that all input/output files live in. By default the current R working directory is used.}
}
\description{
Use plink to fit models. If covariates are to be used assumes that pheno and covar
 files exist and are in the form ofile_pheno.txt and ofile_covar.txt. If coxph is
 to be used assumes ofile_Rplink.R exists and the system is unix
}

