% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plinktoR}
\alias{plinktoR}
\title{Extract data from plink bfile into a Data Frame}
\usage{
plinktoR(ifile, ofile = NULL, snps = NULL, pd = "", wd = "")
}
\arguments{
\item{ifile}{String corresponding to the name of the plink bfile (no extension)}

\item{ofile}{String corresponding to the name of the resulting file which can be read into a dataframe.
It will be called ofile.raw. By default ofile=ifile}

\item{snps}{A vector of snps that will be in the dataframe. By default all snps will be selected}

\item{pd}{directory of plink. Use only if plink is not in your path}

\item{wd}{directory that all files are located in. Is by default set to the current R working directory}
}
\description{
Extracts data from a plink bfile into a Data Frame for use in R.
Can extract all data, or a subset of specified snps
}

