% Generated by roxygen2 (4.0.0): do not edit by hand
\name{qq}
\alias{qq}
\title{Create a qq plot of genetic analysis p-values}
\usage{
qq(ifile, title = "qq plot of p-values", ...)
}
\arguments{
\item{ifile}{the name of the output file of the genetic analysis}

\item{title}{title of plot}

\item{...}{arguments to go into plot}
}
\description{
Create a qq plot of genetic analysis p-values
}

