% Generated by roxygen2 (4.0.0): do not edit by hand
\name{remove_relatives}
\alias{remove_relatives}
\title{Remove relatives}
\usage{
remove_relatives(cutoff = 0.25)
}
\arguments{
\item{cutoff}{numeric corresponding to cutoff ibs.
Default is 0.25.}
}
\description{
Removes the larger ID number for all pairs of individual over
specified IBS
}
\keyword{GWAS}

