% Generated by roxygen2 (4.0.0): do not edit by hand
\name{remove_sex_problems}
\alias{remove_sex_problems}
\title{Remove people with sex problems}
\usage{
remove_sex_problems(pd = "")
}
\arguments{
\item{pd}{directory where plink is. Do not need to specify if plink is in your path
or working directory}
}
\description{
Remove people with sex problems.
}
\keyword{GWAS}

