% Generated by roxygen2 (4.0.0): do not edit by hand
\name{snpinfo}
\alias{snpinfo}
\title{Returns a dataframe with descriptive information about each snp}
\usage{
snpinfo(ifile, pd, wd)
}
\arguments{
\item{ifile}{String corresponding to the name of the plink bfile you want to summarize (no extension)}

\item{pd}{directory of plink. Use only if plink is not in your path}

\item{wd}{directory that all files are located in. Is by default set to the current R working directory}
}
\description{
Returns a dataframe with descriptive information about each snp
}

