% Generated by roxygen2 (4.0.0): do not edit by hand
\name{topsnps}
\alias{topsnps}
\title{Summarize top snp information}
\usage{
topsnps(ifile, ofile = NULL, topn = 0, topprop = 0, topcut = 0,
  wd = "")
}
\arguments{
\item{ifile}{String corresponding to the name of the results file you want to get the top snps from}

\item{ofile}{String corresponding to the name of the resulting files (by default same as ifile)}

\item{topn}{Specify how many snps you want to retreive}

\item{topprop}{specifiy what proportion of snps you want to retreive}

\item{topcut}{specify the p-value cutoff you want to retreive}

\item{wd}{directory that all files are located in. Is by default set to the current R working directory}
}
\description{
Summarize top snp information. Can get top n snps, a proportion of top snps, or all snps with p-values
less than a specified cutoff. Each option will write to a unique file
}

