% Generated by roxygen2 (4.0.0): do not edit by hand
\name{GENfit}
\alias{GENfit}
\title{Fit genetic models}
\usage{
GENfit(id, outcome, covariates = NULL, strata = NULL, type,
  kind = "additive", ifile = "GENmatic", ofile = "GENmatic", qq = F,
  manhattan = F, topn = 10, topprop = 0, topcut = 0, pd = "",
  wd = "")
}
\arguments{
\item{id}{dataframe with 2 columns corresponding to the id in the .fam plink file}

\item{outcome}{dataframe/vector of outcomes. If coxph model first column is time, second column is status}

\item{covariates}{dataframe of covaraites. Default is NULL for no covariates}

\item{strata}{vector of statification membership status. Works only with coxph.}

\item{type}{Type of analysis to do. Must be one of "linear", "logistic", "coxph".}

\item{kind}{string corresponding to additive/dominant/recessive model. Currently only
coxph can do a non-additive analysis}

\item{ifile}{String corresponding to the name of bed/bim/fam plink files. Default is GENmatic}

\item{ofile}{String corresponding to the name of the resulting files. Default is GENmatic}

\item{qq}{Boolean indicating if a QQ plot of the p-values will be created. Default is F}

\item{manhattan}{Boolean indicating if a manhattan plot of the p-values will be created. Default is F}

\item{topn}{If not equal to 0, will create a file with the top n number of snps. Default is 10.}

\item{topprop}{If not equal to 0, will create a file with the top n proportion of snps (0<=n<=1)}

\item{topcut}{If not equal to 0, will create a file with all snps with a p-value <= n}

\item{pd}{Directory that plink is located in. Only specify if it is not in your path}

\item{wd}{Directory that all input/output files live in. By default the current R working directory is used.}
}
\description{
Fit genetic models. Currently can only fit coxph models
}

