% Generated by roxygen2 (4.0.0): do not edit by hand
\name{makeRplink}
\alias{makeRplink}
\title{Make Rplink.txt file}
\usage{
makeRplink(outcome, covariates = NULL, strata = NULL, kind = "additive",
  wd = "", cname = "covar")
}
\arguments{
\item{outcome}{A Vector or Matrix/Dataframe corresponding to the outcome. If survival outcome then
the first column is the time, and the second column is the status.}

\item{covariates}{vector/Dataframe corresponding to confounding factors. Default
(no covariates) is NULL}

\item{strata}{vector corresponding to stratification variable. This is only for coxph models. Default
(no strata) is NULL.}

\item{kind}{additive/dominant/recessive}

\item{wd}{string corresponding to the directory you want to output the
Rplink.R file to. Default is to the current working directory}

\item{cname}{the name of the plink covariate file (no extension) default is covar}
}
\description{
Makes Rplink.txt file for fash coxph gwas.
Any element of the outcome,strata, or null vectors will correspond to either
the column index (not including the 2 id columns) or column name of the
corresponding covaraite in the covar.txt file. Any element of these vectors will
be treated as an index if it consists of the digits 0-9 and will be considered
a column name otherwise. Even if you input an element as "101" (with quotes) it will
be treated as an index to the 101th column not the column name 101. Only one stratification
variable can be used. If you want to use more than one, create a variable that incorporates all of
the strata in one column.
}

